// src/services/wallet-state-core/generated/DsessView.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DsessConnectionType } from "./DsessConnectionType";
import type { DsessDappInfo } from "./DsessDappInfo";
import type { DsessError } from "./DsessError";
import type { DsessRemoteInjectSession } from "./DsessRemoteInjectSession";
import type { DsessStatus } from "./DsessStatus";

export type DsessView = { status: DsessStatus, error: DsessError | null, session: DsessRemoteInjectSession | null, dapp_info: DsessDappInfo | null, connection_type: DsessConnectionType | null, pending_fingerprint: string | null, reconnect_stuck: boolean, chain_id: number, };